:recycle: (pc-13348) rename useCulturalSurvey file to be more generic
