+++ src/features/culturalSurvey/__mocks__/useCulturalSurvey.ts
@@ -2,7 +2,7 @@ import { UseQueryResult } from 'react-query'
 
 import { CulturalSurveyQuestionsResponse } from 'api/gen'
 import { mockedCulturalSurveyQuestions } from 'features/culturalSurvey/fixtures/mockedCulturalSurveyQuestions'
-import { useCulturalSurveyQuestions as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/useCulturalSurveyQuestions'
+import { useCulturalSurveyQuestions as actualUseCulturalSurveyQuestions } from 'features/culturalSurvey/useCulturalSurvey'
 
 export const useCulturalSurveyQuestions: typeof actualUseCulturalSurveyQuestions = jest.fn(
   () =>
