+++ src/features/culturalSurvey/context/CulturalSurveyContextProvider.tsx
@@ -6,7 +6,7 @@ import {
 } from 'features/culturalSurvey/context/reducer'
 import { Action, CulturalSurveyState } from 'features/culturalSurvey/context/types'
 import { createInitialQuestionsList } from 'features/culturalSurvey/helpers/utils'
-import { useCulturalSurveyQuestions } from 'features/culturalSurvey/useCulturalSurveyQuestions'
+import { useCulturalSurveyQuestions } from 'features/culturalSurvey/useCulturalSurvey'
 
 export interface ICulturalSurveyContext extends CulturalSurveyState {
   dispatch: React.Dispatch<Action>
